Updated utility with new class that handles return of multiple time-steps. Updated autotest and example notebook. Added new .cbc for the freyberg mutlilayer transient model...budgets were not saved for recharge or wells in the existing cell-budget file. Model was rerun and resulting .cbc file (freyberg_mlt.cbc) was copied to the example data directory.
