+++ autotest/t038_test.py
@@ -3,25 +3,16 @@ Test zonbud utility
 """
 import os
 import numpy as np
-from flopy.utils import ZoneBudget, read_zbarray, write_zbarray, MfListBudget
+from flopy.utils import CellBudgetFile, ZoneBudget, read_zbarray, write_zbarray
 
 loadpth = os.path.join('..', 'examples', 'data', 'zonbud_examples')
 outpth = os.path.join('temp', 't038')
-cbc_f = 'freyberg.gitcbc'
 
 if not os.path.isdir(outpth):
     os.makedirs(outpth)
 
-# nrow, ncol = 40, 20
-# zon = np.zeros((1, nrow, ncol), np.int)
-# zon[0, :20, :10] = 1
-# zon[0, :20, 10:] = 2
-# zon[0, 20:, :10] = 3
-# zon[0, 20:, 10:] = 4
-zon = np.random.randint(1, 6, size=(1, 40, 20))
 
-
-def test_compare_results_2_zonebudget(rtol=1e-5):
+def test_compare2zonebudget(rtol=1e-2):
     """
     Compares output from zonbud.exe to the budget calculated by the zonbud utility
     using the single-layer freyberg model.
@@ -53,27 +44,23 @@ def test_compare_results_2_zonebudget(rtol=1e-5):
         elif 'Total' in items[0] or 'IN-OUT' in items[0] or 'Percent Error' in items[0]:
             continue
         record = '_'.join(items[0].strip().split()) + '_{}'.format(flow_dir)
+        if record.startswith(('FROM_', 'TO_')):
+            record = '_'.join(record.split('_')[1:])
         vals = [float(i) for i in items[1:-1]]
-        row = (flow_dir, record) + tuple(v for v in vals)
+        row = (record,) + tuple(v for v in vals)
         rows.append(row)
-    dtype_list = [('flow_dir', '<U3'), ('record', '<U50')] + [(z, '<f8') for z in zonenames]
+    dtype_list = [('record', '<U50')] + [(z, '<f8') for z in zonenames]
     dtype = np.dtype(dtype_list)
     zonebudget_recarray = np.array(rows, dtype=dtype)
 
     zon = read_zbarray(os.path.join(loadpth, 'zonef'))
-    cbc_fname = os.path.join(loadpth, 'freyberg.gitcbc')
-    zbud = ZoneBudget(cbc_fname, zon, kstpkper=(kstp, kper))
-    zbudutil_recarray = zbud.recordarray
+    cbc_fname = os.path.join(loadpth, 'freyberg.cbc')
+    zb = ZoneBudget(cbc_fname, zon, kstpkper=(kstp, kper))
+    zbudutil_recarray = zb.get_budget()[0]
 
-    flow_dirs = zonebudget_recarray['flow_dir']
-    records = zonebudget_recarray['record']
-
-    for flowdir, recname in zip(flow_dirs, records):
-        r1 = np.where((zbudutil_recarray['flow_dir'] == flowdir) &
-                      (zbudutil_recarray['record'] == recname))
-        r2 = np.where((zonebudget_recarray['flow_dir'] == flowdir) &
-                      (zonebudget_recarray['record'] == recname))
-
+    for recname in zonebudget_recarray['record']:
+        r1 = np.where((zbudutil_recarray['record'] == recname))
+        r2 = np.where((zonebudget_recarray['record'] == recname))
         a1 = np.array([v for v in zbudutil_recarray[zonenames][r1[0]][0]])
         a2 = np.array([v for v in zonebudget_recarray[zonenames][r2[0]][0]])
         assert a1.shape == a2.shape, 'Array shapes do not match.'
@@ -82,7 +69,7 @@ def test_compare_results_2_zonebudget(rtol=1e-5):
     return
 
 
-def test_compare_mlt_results_2_zonebudget(rtol=1e-5):
+def test_compare2zonebudget_mlt(rtol=1e-2):
     """
     Compares output from zonbud.exe to the budget calculated by the zonbud utility
     using the multilayer transient freyberg model.
@@ -101,97 +88,76 @@ def test_compare_mlt_results_2_zonebudget(rtol=1e-5):
             zonenames = [i.strip() for i in items[1:-1]]
             zonenames = ['_'.join(z.split()) for z in zonenames]
             continue
-        elif 'IN' in items[1]:
+        elif items[1].strip() == 'IN':
             flow_dir = 'IN'
             continue
-        elif 'OUT' in items[1]:
+        elif items[1].strip() == 'OUT':
             flow_dir = 'OUT'
             continue
-        elif 'Total' in items[0] or 'IN-OUT' in items[0] or 'Percent Error' in items[0]:
-            continue
+        elif 'Total' in items[0] or 'IN-OUT' in items[0]:
+            record = '_'.join(items[0].strip().upper().split())
+        elif 'Percent Error' in items[0]:
+            record = 'PERCENT_DISCREPANCY'
         elif items[0] == '' and items[1] == '\n':
             break
-        elif 'Total' in items[0] or 'IN-OUT' in items[0] or 'Percent Error' in items[0]:
-            continue
-        record = '_'.join(items[0].strip().split()) + '_{}'.format(flow_dir)
+        else:
+            record = '_'.join(items[0].strip().split()) + '_{}'.format(flow_dir)
+            if record.startswith(('FROM_', 'TO_')):
+                record = '_'.join(record.split('_')[1:])
         vals = [float(i) for i in items[1:-1]]
-        row = (flow_dir, record) + tuple(v for v in vals)
+        row = (record,) + tuple(v for v in vals)
         rows.append(row)
-    dtype_list = [('flow_dir', '<U3'), ('record', '<U50')] + [(z, '<f8') for z in zonenames]
+    dtype_list = [('record', '<U50')] + [(z, '<f8') for z in zonenames]
     dtype = np.dtype(dtype_list)
     zonebudget_recarray = np.array(rows, dtype=dtype)
 
     zon = read_zbarray(os.path.join(loadpth, 'zonef_mlt'))
-    cbc_fname = os.path.join('..', 'examples', 'data', 'freyberg_multilayer_transient', 'freyberg.cbc')
-    zbud = ZoneBudget(cbc_fname, zon, kstpkper=(kstp, kper))
-    zbudutil_recarray = zbud.recordarray
+    cbc_fname = os.path.join(loadpth, 'freyberg_mlt.cbc')
+    zb = ZoneBudget(cbc_fname, zon, kstpkper=(kstp, kper))
+    zbudutil_recarray = zb.get_budget()[0]
 
-    flow_dirs = zonebudget_recarray['flow_dir']
-    records = zonebudget_recarray['record']
-
-    for flowdir, recname in zip(flow_dirs, records):
-        r1 = np.where((zbudutil_recarray['flow_dir'] == flowdir) &
-                      (zbudutil_recarray['record'] == recname))
-        r2 = np.where((zonebudget_recarray['flow_dir'] == flowdir) &
-                      (zonebudget_recarray['record'] == recname))
+    for recname in zonebudget_recarray['record']:
+        if recname in ['IN-OUT', 'PERCENT_DISCREPANCY']:
+            # Skip these, may not match due to presicion
+            # of the zonebudget output file
+            continue
+        r1 = np.where((zbudutil_recarray['record'] == recname))
+        r2 = np.where((zonebudget_recarray['record'] == recname))
         a1 = np.array([v for v in zbudutil_recarray[zonenames][r1[0]][0]])
         a2 = np.array([v for v in zonebudget_recarray[zonenames][r2[0]][0]])
-        assert a1.shape == a2.shape, 'Array shapes do not match.'
+        assert a1.shape == a2.shape, 'Array shapes do not match ({}).'.format(recname)
         isclose = np.allclose(a1, a2, rtol)
-        assert isclose, 'Zonebudget arrays do not match within a tolerance of {}.'.format(rtol)
+        assert isclose, 'Zonebudget arrays do not match within a tolerance of {} {}.'.format(rtol, recname)
     return
 
 
-def test_zonbud_write_csv_kstpkper():
-    zbud = ZoneBudget(os.path.join(loadpth, cbc_f), zon, kstpkper=(0, 0))
-    zbud.to_csv(os.path.join(outpth, 'zbud_zonbud.csv'), write_format='zonbud')
-    zbud.to_csv(os.path.join(outpth, 'zbud_pandas.csv'), write_format='pandas')
-    return
-
-
-def test_zonbud_write_csv_totim():
-    zbud = ZoneBudget(os.path.join(loadpth, cbc_f), zon, totim=10.)
-    zbud.to_csv(os.path.join(outpth, 'zbud_zonbud.csv'), write_format='zonbud')
-    zbud.to_csv(os.path.join(outpth, 'zbud_pandas.csv'), write_format='pandas')
-    return
-
-
-def test_zonbud_budget():
-    zbud = ZoneBudget(os.path.join(loadpth, cbc_f), zon, kstpkper=(0, 0))
-    assert zbud.get_records().shape[0] > 0, 'No records returned.'
-    recordlist = ['CONSTANT_HEAD_IN', 'FROM_ZONE_1_IN']
-    assert zbud.get_records(recordlist=recordlist, zones=[1, 3]).shape[0] > 0, 'No records returned.'
-    return
-
-
-def test_zonbud_mass_balance():
-    zbud = ZoneBudget(os.path.join(loadpth, cbc_f), zon, kstpkper=(0, 0))
-    assert zbud.get_mass_balance().shape[0] > 0, 'No records returned.'
-    assert zbud.get_total_outflow(zones=3).shape[0] > 0, 'No records returned.'
-    assert zbud.get_total_inflow(zones=(1, 2)).shape[0] > 0, 'No records returned.'
-    assert zbud.get_percent_error().shape[0] > 0, 'No records returned.'
+def test_zonbud_get_budget():
+    cbc_f = os.path.join(loadpth, 'freyberg_mlt.cbc')
+    cbc = CellBudgetFile(cbc_f)
+    zon = read_zbarray(os.path.join(loadpth, 'zonef_mlt'))
+    zb = ZoneBudget(cbc_f, zon)
+    s = 'Number of records returned does not match the number requested.'
+    assert len(zb.get_budget()) == len(cbc.get_kstpkper()), s
     return
 
 
 def test_zonbud_aliases():
-    aliases = {1: 'Trey', 2: 'Mike', 4: 'Page', 0: 'Carini'}
-    zbud = ZoneBudget(os.path.join(loadpth, cbc_f), zon, kstpkper=(0, 0), aliases=aliases)
-    zbud.to_csv(os.path.join(outpth, 'zbud_aliases.csv'), write_format='zonbud')
-    assert zbud.get_records().shape[0] > 0, 'No records returned.'
-    recordlist = ['FROM_Mike_IN']
-    assert zbud.get_records(recordlist=recordlist, zones=['Trey', 3]).shape[0] > 0, 'No records returned.'
+    cbc_f = 'freyberg.cbc'
+    zon = read_zbarray(os.path.join(loadpth, 'zonef'))
+    aliases = {1: 'Trey', 2: 'Mike', 4: 'Wilson', 0: 'Carini'}
+    zb = ZoneBudget(os.path.join(loadpth, cbc_f), zon, kstpkper=(0, 0), aliases=aliases)
+    bud = zb.get_budget()[0]
+    m = bud['record'] == 'Mike_IN'
+    assert bud[m].shape[0] > 0, 'No records returned.'
     return
 
 
-def test_zonbud_mult():
+def test_zonbud_copy():
+    cbc_f = 'freyberg.cbc'
+    zon = read_zbarray(os.path.join(loadpth, 'zonef'))
     cfd = ZoneBudget(os.path.join(loadpth, cbc_f), zon, kstpkper=(0, 0))
-    cfd.to_csv(os.path.join(outpth, 'zbud_zonbud.csv'))
-    mgd = cfd*(7.48052/1000000)
-    mgd.to_csv(os.path.join(outpth, 'zbud_zonbud.csv'))
     cfd2 = cfd.copy()
-    assert cfd.recordarray is not cfd2.recordarray, 'Copied object is a shallow copy.'
-    cfd2.to_csv(os.path.join(outpth, 'zbud_zonbud.csv'))
-    cfd2 / 5
+    assert cfd is not cfd2, 'Copied object is a shallow copy.'
     return
 
 
@@ -204,23 +170,40 @@ def test_zonbud_readwrite_zbarray():
     return
 
 
-def junk():
-    listf = os.path.join('..', 'examples', 'data', 'freyberg_multilayer_transient', 'freyberg.list')
-    bud = MfListBudget(listf)
-    # print(bud.get_record_names())
-    # print(bud.get_kstpkper())
-    print(bud.get_budget(names=['CONSTANT_HEAD_IN']))
-    # print(bud.get_data(kstpkper=None))
-    return
+# def junk():
+#     from flopy.utils import MfListBudget
+#     listf = os.path.join('..', 'examples', 'data', 'freyberg_multilayer_transient', 'freyberg.list')
+#     bud = MfListBudget(listf)
+#     # print(bud.get_kstpkper())
+#     inc, cum = bud.get_budget()
+#     print(repr(inc))
+#     # print(bud.get_data(kstpkper=None))
+#
+#     # zon = read_zbarray(os.path.join(loadpth, 'zonef_mlt'))
+#     # cbc_fname = os.path.join('..', 'examples', 'data', 'freyberg_multilayer_transient', 'freyberg.cbc')
+#     # zbud = ZoneBudget(cbc_fname, zon, kstpkper=(0, 1096))
+#     # print(zbud.get_budget(recordlist=['CONSTANT_HEAD_IN']))
+#     # print(zbud.get_budget(recordlist=['CONSTANT_HEAD_IN']).sum())
+#     return
+#
+# def test_zonbud2():
+#     zon = read_zbarray(os.path.join(loadpth, 'zonef_mlt'))
+#     # zon = np.ones((3, 40, 20), np.int)
+#     cbc_fname = os.path.join(loadpth, 'freyberg.cbc')
+#     bud = ZoneBudget(cbc_fname, zon, totim=[1095., 1096., 1097.])
+#     # print(len(bud))
+#     a = bud.get_budget()[0]
+#     # for i in range(a.shape[0]):
+#     #     print(i, list(a[:][i]))
+#     bud.to_csv(os.path.join(outpth, 'text.csv'))
+#     mul = bud*-1
+#     # print(repr(mul.get_budget()[-1]))
+#     return
 
 if __name__ == '__main__':
-    # junk()
-    test_compare_results_2_zonebudget()
-    test_compare_mlt_results_2_zonebudget()
-    test_zonbud_write_csv_kstpkper()
-    test_zonbud_write_csv_totim()
+    test_compare2zonebudget()
+    test_compare2zonebudget_mlt()
     test_zonbud_aliases()
-    test_zonbud_budget()
-    test_zonbud_mass_balance()
-    test_zonbud_mult()
+    test_zonbud_get_budget()
+    test_zonbud_copy()
     test_zonbud_readwrite_zbarray()
