+++ examples/data/zonbud_examples/freyberg_mlt_zonebudget.csv
@@ -1,21 +1,23 @@
 Time Step,   1,Stress Period,1097,Sim. Time, 1.097000E+03,,
-                ,  ZONE   1      ,  ZONE   2      ,  ZONE   3      ,  ZONE   4      ,  ZONE   5      ,
-                ,    IN          ,    IN          ,    IN          ,    IN          ,    IN          ,
-   CONSTANT HEAD,    0.000000E+00,    1.723660E+02,    0.000000E+00,    0.000000E+00,    1.454174E+02,
-   FROM ZONE   1,    0.000000E+00,    1.894527E+03,    2.901152E+03,    2.183301E+03,    2.755434E+03,
-   FROM ZONE   2,    2.256005E+03,    0.000000E+00,    2.086318E+03,    2.016456E+03,    2.533748E+03,
-   FROM ZONE   3,    1.978395E+03,    2.344866E+03,    0.000000E+00,    2.800115E+03,    2.451944E+03,
-   FROM ZONE   4,    1.811187E+03,    2.004503E+03,    2.940716E+03,    0.000000E+00,    2.731191E+03,
-   FROM ZONE   5,    2.950723E+03,    3.400900E+03,    2.435741E+03,    2.458422E+03,    0.000000E+00,
-Total IN        ,    8.996310E+03,    9.817163E+03,    1.036393E+04,    9.458294E+03,    1.061773E+04,
-                ,   OUT          ,   OUT          ,   OUT          ,   OUT          ,   OUT          ,
-   CONSTANT HEAD,    1.954154E+02,    2.139180E+02,    2.561286E+01,    1.829173E+02,    1.274998E+02,
-     TO ZONE   1,    0.000000E+00,    2.256005E+03,    1.978395E+03,    1.811187E+03,    2.950723E+03,
-     TO ZONE   2,    1.894527E+03,    0.000000E+00,    2.344866E+03,    2.004503E+03,    3.400900E+03,
-     TO ZONE   3,    2.901152E+03,    2.086318E+03,    0.000000E+00,    2.940716E+03,    2.435741E+03,
-     TO ZONE   4,    2.183301E+03,    2.016456E+03,    2.800115E+03,    0.000000E+00,    2.458422E+03,
-     TO ZONE   5,    2.755434E+03,    2.533748E+03,    2.451944E+03,    2.731191E+03,    0.000000E+00,
-Total OUT       ,    9.929830E+03,    9.106445E+03,    9.600933E+03,    9.670515E+03,    1.137329E+04,
- IN-OUT         ,   -9.335193E+02,    7.107179E+02,    7.629944E+02,   -2.122211E+02,   -7.555520E+02,
-Percent Error   ,   -9.864867E+00,    7.511442E+00,    7.643374E+00,   -2.218864E+00,   -6.871459E+00,
+                ,  ZONE   1      ,  ZONE   2      ,  ZONE   3      ,
+                ,    IN          ,    IN          ,    IN          ,
+   CONSTANT HEAD,    0.000000E+00,    0.000000E+00,    8.614904E+01,
+           WELLS,    0.000000E+00,    0.000000E+00,    0.000000E+00,
+        RECHARGE,    1.306933E+03,    1.299465E+03,    1.277060E+03,
+   FROM ZONE   0,    4.802645E+03,    3.535905E+03,    3.882709E+03,
+   FROM ZONE   1,    0.000000E+00,    3.485947E+03,    3.202062E+03,
+   FROM ZONE   2,    3.823094E+03,    0.000000E+00,    2.978464E+03,
+   FROM ZONE   3,    3.579749E+03,    3.084417E+03,    0.000000E+00,
+Total IN        ,    1.351242E+04,    1.140573E+04,    1.142644E+04,
+                ,   OUT          ,   OUT          ,   OUT          ,
+   CONSTANT HEAD,    1.015419E+02,    2.324631E+02,    2.043376E+02,
+           WELLS,    2.656800E+03,    0.000000E+00,    0.000000E+00,
+        RECHARGE,    0.000000E+00,    0.000000E+00,    0.000000E+00,
+     TO ZONE   0,    4.066082E+03,    4.371712E+03,    4.557932E+03,
+     TO ZONE   1,    0.000000E+00,    3.823094E+03,    3.579749E+03,
+     TO ZONE   2,    3.485947E+03,    0.000000E+00,    3.084417E+03,
+     TO ZONE   3,    3.202062E+03,    2.978464E+03,    0.000000E+00,
+Total OUT       ,    1.351243E+04,    1.140573E+04,    1.142644E+04,
+ IN-OUT         ,   -1.238740E-02,    1.146679E-03,    8.515302E-03,
+Percent Error   ,   -9.167415E-05,    1.005353E-05,    7.452280E-05,
 ,
