+++ examples/data/zonbud_examples/freyberg_zonebudget.csv
@@ -1,27 +1,25 @@
 Time Step,   1,Stress Period,   1,Sim. Time, 1.000000E+01,,
-                ,  ZONE   1      ,  ZONE   2      ,  ZONE   3      ,  ZONE   4      ,  ZONE   5      ,
-                ,    IN          ,    IN          ,    IN          ,    IN          ,    IN          ,
-   CONSTANT HEAD,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,
-           WELLS,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,
-   RIVER LEAKAGE,    4.194190E-03,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,
-        RECHARGE,    1.610000E-02,    1.110000E-02,    1.410000E-02,    1.450000E-02,    1.370000E-02,
-   FROM ZONE   1,    0.000000E+00,    1.608695E-02,    2.827535E-02,    2.821370E-02,    2.368117E-02,
-   FROM ZONE   2,    2.216422E-02,    0.000000E+00,    1.817669E-02,    1.397196E-02,    1.593054E-02,
-   FROM ZONE   3,    2.683798E-02,    1.887462E-02,    0.000000E+00,    2.265548E-02,    2.222089E-02,
-   FROM ZONE   4,    2.692761E-02,    1.493824E-02,    2.405315E-02,    0.000000E+00,    1.766146E-02,
-   FROM ZONE   5,    2.650290E-02,    1.268903E-02,    1.733073E-02,    2.343331E-02,    0.000000E+00,
-Total IN        ,    1.227269E-01,    7.368884E-02,    1.019359E-01,    1.027745E-01,    9.319406E-02,
-                ,   OUT          ,   OUT          ,   OUT          ,   OUT          ,   OUT          ,
-   CONSTANT HEAD,    1.389692E-03,    4.085897E-05,    6.519382E-04,    2.263298E-03,    3.894946E-04,
-           WELLS,    1.302000E-02,    0.000000E+00,    0.000000E+00,    5.130000E-03,    3.900000E-03,
-   RIVER LEAKAGE,    1.206024E-02,    3.404616E-03,    1.069514E-02,    1.180087E-02,    8.948695E-03,
-        RECHARGE,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,    0.000000E+00,
-     TO ZONE   1,    0.000000E+00,    2.216422E-02,    2.683798E-02,    2.692761E-02,    2.650290E-02,
-     TO ZONE   2,    1.608695E-02,    0.000000E+00,    1.887462E-02,    1.493824E-02,    1.268903E-02,
-     TO ZONE   3,    2.827535E-02,    1.817669E-02,    0.000000E+00,    2.405315E-02,    1.733073E-02,
-     TO ZONE   4,    2.821370E-02,    1.397196E-02,    2.265548E-02,    0.000000E+00,    2.343331E-02,
-     TO ZONE   5,    2.368117E-02,    1.593054E-02,    2.222089E-02,    1.766146E-02,    0.000000E+00,
-Total OUT       ,    1.227271E-01,    7.368889E-02,    1.019360E-01,    1.027746E-01,    9.319417E-02,
- IN-OUT         ,   -2.004487E-07,   -5.371930E-08,   -1.257563E-07,   -1.649369E-07,   -1.032776E-07,
-Percent Error   ,   -1.633289E-04,   -7.290017E-05,   -1.233679E-04,   -1.604842E-04,   -1.108198E-04,
+                ,  ZONE   1      ,  ZONE   2      ,  ZONE   3      ,
+                ,    IN          ,    IN          ,    IN          ,
+   CONSTANT HEAD,    0.000000E+00,    0.000000E+00,    0.000000E+00,
+           WELLS,    0.000000E+00,    0.000000E+00,    0.000000E+00,
+   RIVER LEAKAGE,    4.194190E-03,    0.000000E+00,    0.000000E+00,
+        RECHARGE,    1.970000E-02,    1.460000E-02,    1.680000E-02,
+   FROM ZONE   0,    3.600075E-02,    3.503814E-02,    3.750513E-02,
+   FROM ZONE   1,    0.000000E+00,    2.955931E-02,    4.451611E-02,
+   FROM ZONE   2,    2.813982E-02,    0.000000E+00,    2.709453E-02,
+   FROM ZONE   3,    4.001517E-02,    1.918741E-02,    0.000000E+00,
+Total IN        ,    1.280499E-01,    9.838486E-02,    1.259158E-01,
+                ,   OUT          ,   OUT          ,   OUT          ,
+   CONSTANT HEAD,    2.908484E-03,    6.519382E-04,    7.445060E-04,
+           WELLS,    0.000000E+00,    4.300000E-03,    1.775000E-02,
+   RIVER LEAKAGE,    1.465203E-02,    1.016300E-02,    1.201904E-02,
+        RECHARGE,    0.000000E+00,    0.000000E+00,    0.000000E+00,
+     TO ZONE   0,    3.641424E-02,    2.803571E-02,    3.619980E-02,
+     TO ZONE   1,    0.000000E+00,    2.813982E-02,    4.001517E-02,
+     TO ZONE   2,    2.955931E-02,    0.000000E+00,    1.918741E-02,
+     TO ZONE   3,    4.451611E-02,    2.709453E-02,    0.000000E+00,
+Total OUT       ,    1.280502E-01,    9.838500E-02,    1.259159E-01,
+ IN-OUT         ,   -2.355989E-07,   -1.347317E-07,   -1.505664E-07,
+Percent Error   ,   -1.839897E-04,   -1.369435E-04,   -1.195770E-04,
 ,
